Feat: Record 수정
