--- Escaper/Escaper/Infrastructure/DataMapping/RecordInfoDTO.swift
@@ -1,38 +1,0 @@
-//
-//  RecordInfoDTO.swift
-//  Escaper
-//
-//  Created by 최완식 on 2021/11/10.
-//
-
-import Foundation
-
-struct RecordInfoDTO: Codable {
-    var imageUrlString: String
-    var userEmail: String
-    var roomId: String
-    var satisfaction: Double
-    var isSuccess: Bool
-    var time: Int
-
-    func toDictionary() -> [String: Any] {
-        let dictionary: [String: Any] = [
-            "imageUrlString": self.imageUrlString,
-            "userEmail": self.userEmail,
-            "roomId": self.roomId,
-            "satisfaction": self.satisfaction,
-            "isSuccess": self.isSuccess,
-            "time": self.time
-        ]
-        return dictionary
-    }
-
-    func toDomain() -> RecordInfo {
-        return RecordInfo(imageUrlString: self.imageUrlString,
-                          userEmail: self.userEmail,
-                          roomId: self.roomId,
-                          satisfaction: Rating(rawValue: Int(self.satisfaction.rounded())) ?? Rating.zero,
-                          isSuccess: self.isSuccess,
-                          time: self.time)
-    }
-}
